+++ observer_pattern/gulpfile.js
@@ -1,5 +1,7 @@
-var gulp = require("gulp");
-var uglify = require("gulp-uglify");
+var gulp = require("gulp"),
+    babel = require("gulp-babel"),
+    uglify = require("gulp-uglify"),
+    plumber = require('gulp-plumber');
 
 /*
 gulp.task(“タスク名”,function() {});でタスクの登録
@@ -9,11 +11,13 @@ gulp.dest(“出力先”)で出力先に処理を施したファイルを出力
 */
 
 gulp.task("js", function() {
-    gulp.src(["js/**/*.js","!js/min/**/*.js"])
-        .pipe(uglify())
-        .pipe(gulp.dest("./js/min"));
+    gulp.src("js/es6/**/*.js")
+        .pipe(plumber())          // 監視中のエラーによる強制停止を回避
+        .pipe(babel())            // ES6をES5準拠に変換
+        .pipe(uglify())           // minify
+        .pipe(gulp.dest("./js")); //出力
 });
 
 gulp.task("default", function(){
-    gulp.watch(["js/**/*.js", "!js/min/**/*.js"], ["js"]);
+    gulp.watch("js/es6/**/*.js", ["js"]);
 });
