+++ observer_pattern/index.html
@@ -13,11 +13,11 @@
                 <label for="">アカウントを作成</label>
             </div>
             <div class="col">
-                <input type="text" placeholder="4文字以上8文字以内で入力してください" data-minLength="4" data-maxLength="8" required>
+                <input type="text" placeholder="2文字以上10文字以内で入力してください" data-minlength="2" data-maxlength="10" required>
                 <ul>
                     <li data-error="required">必須項目です</li>
-                    <li data-error="minLength">4文字以上で入力してください</li>
-                    <li data-error="maxLength">8文字以内で入力してください</li>
+                    <li data-error="minlength">2文字以上で入力してください</li>
+                    <li data-error="maxlength">10文字以内で入力してください</li>
                 </ul>
             </div>
         </div>
@@ -26,16 +26,16 @@
                 <label for="">パスワードを作成</label>
             </div>
             <div class="col">
-                <input type="text" placeholder="4文字以上8文字以内で入力してください" data-minLength="4" data-maxLength="8" required>
+                <input type="text" placeholder="4文字以上8文字以内で入力してください" data-minlength="4" data-maxlength="8" required>
                 <ul>
                     <li data-error="required">必須項目です</li>
-                    <li data-error="minLength">4文字以上で入力してください</li>
-                    <li data-error="maxLength">8文字以内で入力してください</li>
+                    <li data-error="minlength">4文字以上で入力してください</li>
+                    <li data-error="maxlength">8文字以内で入力してください</li>
                 </ul>
             </div>
         </div>
     </div>
     <script type="text/javascript" src="http://ajax.googleapis.com/ajax/libs/jquery/2.0.3/jquery.min.js"></script>
-    <script type="text/javascript" src="js/min/index.js"></script>
+    <script type="text/javascript" src="js/index.js"></script>
 </body>
 </html>
