+++ observer_pattern/js/index.js
@@ -1,159 +1,1 @@
-/**
- * Viewから値を受け取って、値に対してバリデーションを実行する.
- * バリデーションの結果に応じてイベントを通知する.
- */
-
-$(function(){
-  $("input").each(function() {
-    new AppView(this);
-  });
-});
-
-/**
- * Modelの作成
- */
-var AppModel = (function(){
-
-  function AppModel(attrs) {
-    // new しないとエラーにする.関数呼び出しはNG.
-    this.initialize.apply(this, arguments);
-  }
-
-  AppModel.prototype.initialize = function(attrs) {
-    this.val = "";
-
-    // Validateの定義
-    this.attrs = {
-      required : "",
-      maxLength : 8,
-      minLength : 4
-    };
-
-    // Observerの機構を実装.
-    this.listeners = {
-      valid: [],
-      invalid: []
-    };
-  };
-
-  AppModel.prototype.on = function(event, func) {
-    this.listeners[event].push(func);
-  };
-
-  AppModel.prototype.trigger = function(event) {
-    $.each(this.listeners[event], function(){
-      this();
-    });
-  };
-
-  AppModel.prototype.set = function(val) {
-    if (this.val === val) return;
-    this.val = val;
-    this.validate();
-  };
-
-  AppModel.prototype.validate = function() {
-    var val;
-
-    // チェック結果確認用
-    this.errors = [];
-
-    // attrで定義したチェックを実行
-    for (var key in this.attrs) {
-      val = this.attrs[key];
-      if (!this[key](val)) this.errors.push(key);
-    }
-
-    // Validationの結果に応じた処理を実行
-    this.trigger(!this.errors.length ? "valid" : "invalid");
-  };
-
-  // 必須チェック
-  AppModel.prototype.required = function() {
-    return this.val !== "";
-  };
-
-  // 文字数上限チェック
-  AppModel.prototype.maxLength = function(num) {
-    return this.val.length <= num;
-  };
-
-  // 文字数下限チェック
-  AppModel.prototype.minLength = function(num) {
-    return num <= this.val.length;
-  };
-
-  return AppModel;
-})();
-
-/**
- * Viewの作成
- */
-var AppView = (function(){
-  function AppView() {
-    // new しないとエラーにする.関数呼び出しはNG.
-    this.initialize.apply(this, arguments);
-    this.handleEvents();
-  }
-
-  AppView.prototype.initialize = function(el) {
-
-    // エレメントを取得.
-    this.$el = $(el);
-
-    // エラー表示用のリストを取得.
-    this.$list = this.$el.next().children();
-
-    // エレメントのdata属性の値を取得.
-    var obj = this.$el.data();
-
-    // エレメントにrequired属性がついているか判定
-    if (this.$el.prop("required")) {
-      obj["required"] = "";
-    }
-
-    // Modelを作成し、Viewのプロパティとする.
-    this.model = new AppModel(obj);
-  };
-
-  AppView.prototype.handleEvents = function(){
-    var self = this;
-
-    // DomのKeyupとViewのイベントを関連付ける
-    this.$el.on("keyup", function(e) {
-      self.onKeyUp(e);
-    });
-
-    // ModelとViewのイベントを関連付ける
-    this.model.on("valid", function(){
-      self.onValid();
-    });
-
-    this.model.on("invalid", function(){
-      self.onInValid();
-    });
-  };
-
-  AppView.prototype.onKeyUp = function(e){
-    var $target = $(e.currentTarget);
-    this.model.set($target.val());
-  };
-
-  AppView.prototype.onValid = function() {
-    this.$el.removeClass("error");
-    this.$list.hide();
-  };
-
-  AppView.prototype.onInValid = function() {
-    var self = this;
-    this.$el.addClass("error");
-    this.$list.hide();
-
-    $.each(this.model.errors, function(index, val) {
-      self.$list.filter("[data-error=\"" + val + "\"]").show();
-    });
-  };
-
-  return AppView;
-})();
-
+"use strict";function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}var _createClass=function(){function e(e,t){for(var i=0;i<t.length;i++){var n=t[i];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}return function(t,i,n){return i&&e(t.prototype,i),n&&e(t,n),t}}();$(function(){$("input").each(function(){new AppView(this)})});var AppModel=function(){function e(t){_classCallCheck(this,e),this.val="",this.attrs=t,void 0===t&&(this.attrs={required:"",maxlength:8,minlength:4}),this.listeners={valid:[],invalid:[]}}return _createClass(e,[{key:"on",value:function(e,t){this.listeners[e].push(t)}},{key:"trigger",value:function(e){this.listeners[e].map(function(e){e()})}},{key:"set",value:function(e){this.val!==e&&(this.val=e,this.validate())}},{key:"validate",value:function(){var e=void 0;this.errors=[];for(var t in this.attrs)e=this.attrs[t],this[t](e)||this.errors.push(t);this.trigger(this.errors.length?"invalid":"valid")}},{key:"required",value:function(){return""!==this.val}},{key:"maxlength",value:function(e){return this.val.length<=e}},{key:"minlength",value:function(e){return e<=this.val.length}}]),e}(),AppView=function(){function e(t){_classCallCheck(this,e),this.$el=$(t),this.$list=this.$el.next().children();var i=this.$el.data();this.$el.prop("required")&&(i.required=""),this.model=new AppModel(i),this.handleEvents()}return _createClass(e,[{key:"handleEvents",value:function(){var e=this;this.$el.on("keyup",function(t){e.onKeyUp(t)}),this.model.on("valid",function(){e.onValid()}),this.model.on("invalid",function(){e.onInValid()})}},{key:"onKeyUp",value:function(e){var t=$(e.currentTarget);this.model.set(t.val())}},{key:"onValid",value:function(){this.$el.removeClass("error"),this.$list.hide()}},{key:"onInValid",value:function(){var e=this;this.$el.addClass("error"),this.$list.hide(),this.model.errors.map(function(t){e.$list.filter('[data-error="'+t+'"]').show()})}}]),e}();
