--- observer_pattern/js/min/index.js
@@ -1,1 +1,0 @@
-$(function(){$("input").each(function(){new AppView(this)})});var AppModel=function(){function t(t){this.initialize.apply(this,arguments)}return t.prototype.initialize=function(t){this.val="",this.attrs={required:"",maxLength:8,minLength:4},this.listeners={valid:[],invalid:[]}},t.prototype.on=function(t,i){this.listeners[t].push(i)},t.prototype.trigger=function(t){$.each(this.listeners[t],function(){this()})},t.prototype.set=function(t){this.val!==t&&(this.val=t,this.validate())},t.prototype.validate=function(){var t;this.errors=[];for(var i in this.attrs)t=this.attrs[i],this[i](t)||this.errors.push(i);this.trigger(this.errors.length?"invalid":"valid")},t.prototype.required=function(){return""!==this.val},t.prototype.maxLength=function(t){return this.val.length<=t},t.prototype.minLength=function(t){return t<=this.val.length},t}(),AppView=function(){function t(){this.initialize.apply(this,arguments),this.handleEvents()}return t.prototype.initialize=function(t){this.$el=$(t),this.$list=this.$el.next().children();var i=this.$el.data();this.$el.prop("required")&&(i.required=""),this.model=new AppModel(i)},t.prototype.handleEvents=function(){var t=this;this.$el.on("keyup",function(i){t.onKeyUp(i)}),this.model.on("valid",function(){t.onValid()}),this.model.on("invalid",function(){t.onInValid()})},t.prototype.onKeyUp=function(t){var i=$(t.currentTarget);this.model.set(i.val())},t.prototype.onValid=function(){this.$el.removeClass("error"),this.$list.hide()},t.prototype.onInValid=function(){var t=this;this.$el.addClass("error"),this.$list.hide(),$.each(this.model.errors,function(i,n){t.$list.filter('[data-error="'+n+'"]').show()})},t}();
