+++ observer_pattern/package.json
@@ -1,7 +1,7 @@
 {
   "name": "observer_pattern",
   "version": "1.0.0",
-  "description": "GulpとBabel使ってみるてすと",
+  "description": "GulpとBabel使ってみた",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -19,6 +19,9 @@
   "devDependencies": {
     "gulp": "^3.9.0",
     "gulp-babel": "^5.1.0",
+    "gulp-concat": "^2.5.2",
+    "gulp-notify": "^2.2.0",
+    "gulp-plumber": "^1.0.1",
     "gulp-uglify": "^1.2.0"
   }
 }
